Moved instruction to ks-core

+++ workspace/ks-core/src/compiler_new/compiler.rs
@@ -2,7 +2,6 @@ use std::collections::{BTreeMap, HashMap};
 
 use ks_global::utils::ks_error::KsError;
 use ks_global::utils::ks_result::KsResult;
-use ks_vm_new::{Constant, Instruction, Program};
 
 use crate::parser::expression::Expression;
 use crate::parser::identifier_tail::IdentifierTail;
@@ -11,7 +10,10 @@ use crate::parser::parameter::Parameter;
 use crate::parser::statement::Statement;
 
 use super::collection::Collection;
+use super::constant::Constant;
 use super::environment::Environment;
+use super::instructions::Instruction;
+use super::program::Program;
 use super::slot::Slot;
 use super::types::{CollectionId, NativeId, Pointer, VariableId};
 
@@ -525,14 +527,14 @@ impl CompilerNew {
         if let Some(name) = last_name {
             if let Some(native_id) = self.environment.native_function(name) {
                 let native_id = *native_id;
-                self.insert(Instruction::CallNative(native_id, arguments))?;
+                self.insert(Instruction::CallNative(native_id as u32, arguments as u32))?;
                 return Ok(());
             } else {
                 return Err(KsError::parse("Cannot find nativeId"));
             }
         }
 
-        self.insert(Instruction::Call(arguments))?;
+        self.insert(Instruction::Call(arguments as u32))?;
 
         Ok(())
     }
